requirements.txt
aiofiles==23.2.1
aiogram==3.2.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
certifi==2023.11.17
charset-normalizer==3.3.2
env==0.1.0
environment==1.0.0
environs==9.5.0
frozenlist==1.4.0
idna==3.6
magic-filter==1.0.12
marshmallow==3.20.1
multidict==6.0.4
packaging==23.2
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
pytz==2023.3.post1
typing_extensions==4.7.1
yarl==1.9.4
